move cosmos related cmds under its own subgroup
