+++ src/enums.ts
@@ -1,0 +1,6 @@
+// it's a workaround for ./types.d.ts failing to resolve if the enum is used as
+// value (not only the type)
+export enum SignerType {
+  FIREBLOCKS = 'fireblocks',
+  LOCAL = 'local'
+}
